전처리기 객체 위치를 옮긴다
 - 다른 의존 객체와 동일한 모듈을 갖도록 한다
