+++ README.md
@@ -3,7 +3,7 @@
 
 ### 구현한 것
 - 이력서를 구성하는 각 섹션을 [텍스트 파일](src/)로부터 읽어와 [PDF](output.pdf)로 변환하는 [프로그램](lib/mk_pdf.rb)을 구현
-- 중첩 구조 등 복잡한 로직이 필요한 섹션의 경우, [파서](lib/preproc.rb) 구현
+- 중첩 구조 등 복잡한 로직이 필요한 섹션의 경우, [파서](lib/mk_resume/preproc.rb) 구현
 
 ### 프로그램으로 개선한 것
 - 이력서 레이아웃 조정에 드는 시간이 감소
